+++ application/controllers/Admin.php
@@ -34,7 +34,7 @@ class Admin extends CI_Controller
     {
         $data['title'] = 'Manajemen User';
         $data['user'] = $this->Admin_model->getLoggedUser($this->session->userdata('nip'));
-        $data['user_data'] = $this->Admin_model->getUsersData();
+        // $data['user_data'] = $this->Admin_model->getAllUsers();
 
         helper_log("access", "Mengakses menu Manajemen User");
 
@@ -42,6 +42,13 @@ class Admin extends CI_Controller
         $this->load->view('templates/main_sidebar');
         $this->load->view('admin/v_manajemenUser');
         $this->load->view('templates/main_footer');
+    }
+
+    // AMBIL DATA PEGAWAI
+    public function getAllUsers()
+    {
+        $users = $this->Admin_model->getAllUsers();
+        echo json_encode($users);
     }
 
     //Fungsi tambah pegawai