+++ changeLog.md
@@ -1,4 +1,6 @@
 ### 16 Desember 2019
 - Mengubah routes menjadi lebih ciamik
 - Membuat index untuk admin
-- 
+  
+### 17 Desember 2019
+- AJAX Datatables User (ongoing)